added pw cc option to collapse
